requirements.txt
captcha==0.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
pandas==1.3.5
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
